Fixed compilation issue with old GCC (<=4.4)
Compiling with a version of GCC <= 4.4 results in compilation errors
as most of the time related functions and types, like timespec, are
not available within the C standard library, as they are POSIX
extensions. Using `-std=gnu99' solve the issue.

Additionally, we have move most of the code in the header into a
seperate C file. I guess this means this isn't a 'header only'
solution anymore...